+++ CHANGELOG.md
@@ -33,6 +33,12 @@ const session = await getSessionFromStorage(sessionId, {
 
 The following changes have been implemented but not released yet:
 
+### Bugfix
+
+#### browser and node
+
+- Fixed the usage of client information from previous dynamic registration that have no expiration date.
+
 ## [2.5.0](https://github.com/inrupt/solid-client-authn-js/releases/tag/v2.5.0) - 2025-05-09
 
 ### Feature